Move main formatter code to separate file
